app/backend/requirements.txt
# FastAPI backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication  
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
requests==2.31.0 